requirements.txt
scikit-fuzzy
numpy
matplotlib
streamlit
openai
scipy
networkx 
packaging
python-dotenv
google.generativeai
pytest